Fix test and example imports after refactoring OSDMap types
Update denc tests and examples to import OSDMap, PgPool, PgMergeMeta
from osdclient crate instead of denc. Add osdclient and crush as
dev-dependencies to support the tests.

Changes:
- Add crush and osdclient to denc dev-dependencies
- Update imports in all denc integration tests
- Update example test_osdmap_decode.rs imports
- Fix unnecessary cast in object_placement_test.rs

+++ crates/denc/tests/object_placement_test.rs
@@ -1,5 +1,7 @@
 use bytes::Bytes;
-use denc::{OSDMap, PgId, VersionedEncode};
+use crush::PgId;
+use denc::VersionedEncode;
+use osdclient::OSDMap;
 use std::fs;
 use std::path::PathBuf;
 
@@ -71,7 +73,7 @@ fn test_object_placement_pipeline() {
 
                             // Verify PG is within valid range
                             assert_eq!(
-                                pg.pool, *pool_id as u64,
+                                pg.pool, *pool_id,
                                 "PG pool ID should match input pool ID"
                             );
                             assert!(
